distinct data to domain and domain to data mapper
