+++ feature/profile-edit/impl/src/main/java/ru/yeahub/profile_edit/impl/data/ProfileEditDataToDomainMapper.kt
@@ -5,8 +5,6 @@ import ru.yeahub.network_api.models.GetProfileForUserResponse
 import ru.yeahub.network_api.models.GetSkillResponse
 import ru.yeahub.network_api.models.GetUserProfileResponse
 import ru.yeahub.network_api.models.SocialNetworkDto
-import ru.yeahub.network_api.models.UpdateProfileRequest
-import ru.yeahub.network_api.models.UpdateUserRequest
 import ru.yeahub.profile_edit.impl.domain.models.DomainProfileEditData
 import ru.yeahub.profile_edit.impl.domain.models.DomainProfileEditSkill
 import ru.yeahub.profile_edit.impl.domain.models.DomainProfileEditSocialPlatform
@@ -37,70 +35,12 @@ internal class ProfileEditDataToDomainMapper {
         )
     }
 
-    fun mapToUpdateProfileRequest(
-        profile: DomainProfileEditData,
-        cachedProfile: GetProfileForUserResponse,
-        cachedUser: GetUserProfileResponse,
-        cachedAllSkills: List<GetSkillResponse>,
-        allSpecializations: List<DomainProfileEditSpecialization>,
-    ): UpdateProfileRequest {
-        val skillIds = profile.selectedSkills.mapNotNull { skill ->
-            cachedAllSkills.find { it.title == skill.name }?.id
-        }
-
-        val specializationId = if (profile.specialization != null) {
-            resolveSpecializationId(
-                profile.specialization,
-                allSpecializations,
-            )
-        } else {
-            cachedProfile.specializationId
-        }
-
-        return UpdateProfileRequest(
-            profileType = cachedProfile.profileType,
-            specializationId = specializationId,
-            markingWeight = cachedProfile.markingWeight,
-            description = wrapInHtmlPTags(profile.aboutMe),
-            socialNetwork = mapSocialLinksToDto(profile.socialLinks),
-            imageSrc = cachedProfile.imageSrc,
-            isActive = cachedProfile.isActive,
-            profileSkills = skillIds,
-            user = cachedUser,
-        )
-    }
-
-    fun mapToUpdateUserRequest(
-        profile: DomainProfileEditData,
-        cachedUser: GetUserProfileResponse,
-        avatarBase64: String?,
-        avatarDeleted: Boolean,
-    ): UpdateUserRequest {
-        return UpdateUserRequest(
-            username = profile.nickname,
-            country = cachedUser.country,
-            city = profile.location,
-            birthday = cachedUser.birthday,
-            address = cachedUser.address,
-            avatarUrl = if (avatarDeleted) "" else cachedUser.avatarUrl,
-            avatarImage = avatarBase64,
-        )
-    }
-
     private fun resolveSpecializationName(
         specializationId: Long?,
         specializations: List<DomainProfileEditSpecialization>,
     ): String? {
         if (specializationId == null || specializationId == 0L) return null
         return specializations.find { it.id == specializationId }!!.title
-    }
-
-    private fun resolveSpecializationId(
-        specName: String,
-        allSpecializations: List<DomainProfileEditSpecialization>,
-    ): Long {
-        val id = allSpecializations.find { it.title == specName }!!.id
-        return id
     }
 
     private fun mapSocialNetworkToDomain(
@@ -111,17 +51,6 @@ internal class ProfileEditDataToDomainMapper {
             val platform = codeToPlatform(dto.code) ?: return@mapNotNull null
             platform to dto.title
         }.toMap()
-    }
-
-    private fun mapSocialLinksToDto(
-        socialLinks: Map<DomainProfileEditSocialPlatform, String>,
-    ): List<SocialNetworkDto> {
-        return socialLinks.map { (platform, url) ->
-            SocialNetworkDto(
-                code = platform.name,
-                title = url,
-            )
-        }
     }
 
     fun mapSkillToDomain(skill: GetSkillResponse): DomainProfileEditSkill {
@@ -136,26 +65,7 @@ internal class ProfileEditDataToDomainMapper {
 
         val spanned = Html.fromHtml(html, Html.FROM_HTML_MODE_COMPACT)
 
-        return spanned.toString()
-            .trimEnd()
-            .replace("\u00A0", " ")
-    }
-
-    private fun wrapInHtmlPTags(text: String): String {
-        if (text.isBlank()) return ""
-
-        return text.lines().joinToString(separator = "") { line ->
-            "<p>${line.escapeHtml()}</p>"
-        }
-    }
-
-    private fun String.escapeHtml(): String {
-        return this
-            .replace("&", "&amp;")
-            .replace("<", "&lt;")
-            .replace(">", "&gt;")
-            .replace("\"", "&quot;")
-            .replace("'", "&#39;")
+        return spanned.toString().trimEnd().replace("\u00A0", " ")
     }
 
     private fun codeToPlatform(code: String): DomainProfileEditSocialPlatform? = when (code) {
